requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2023.11.17
channels==4.0.0
charset-normalizer==3.3.2
Django==4.2.8
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environs==11.0.0
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
magic-filter==1.0.12
marshmallow==3.21.1
multidict==6.0.4
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
simplejwt==2.0.1
sqlparse==0.4.4
typing==3.7.4.3
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
yarl==1.9.4
